third_party/WebKit/LayoutTests/http/tests/security/mixedContent/insecure-sync-post-xhr-blocked-expected.txt

ALERT: Executing script in the child window.
ALERT: XHR POST opened.
CONSOLE ERROR: line 7: [blocked] The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-sync-xhr-post.html' was loaded over HTTPS, but ran insecure content from 'http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*&headers=Content-Type': this content should also be loaded over HTTPS.

ALERT: exception! NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php?origin=*&headers=Content-Type'.
This test opens a HTTPS window that loads insecure data via XHR. We should trigger a mixed content callback and block the load.
